Move SimilarityMeasure to server code
